+++ src/SmartHopper.Providers.MistralAI/MistralAISettings.cs
@@ -63,7 +63,7 @@ namespace SmartHopper.Providers.MistralAI
             this.maxTokensNumeric = new NumericUpDown
             {
                 Minimum = 1,
-                Maximum = 4096,
+                Maximum = 100000,
                 Value = 150,
                 Dock = DockStyle.Fill,
             };